update(cherryusb): update to v1.5.2

+++ components/drivers/usb/cherryusb/common/usb_hc.h
@@ -39,7 +39,7 @@ struct usbh_urb {
     struct usbh_hubport *hport;
     struct usb_endpoint_descriptor *ep;
     uint8_t data_toggle;
-    uint8_t interval;
+    uint32_t interval;
     struct usb_setup_packet *setup;
     uint8_t *transfer_buffer;
     uint32_t transfer_buffer_length;
